// src/types/screens/charactersType.ts
export type PersonajesType = {
    nacimiento_ano: string
    creado: string,
    editado: string
    ojo_color: string
    genero: string
    cabello_color: string
    altura: string
    hogarplaneta: string
    masa: string
    nombre: string
    piel_color: string
    url: string
    especies: string[]
    naves: string[]
    vehiculos: string[]
    peliculas: string[]
}

export type PersonajesInglesType = {
    birth_year: string
    created: string,
    edited: string
    eye_color: string
    gender: string
    hair_color: string
    height: string
    homeworld: string
    mass: string
    name: string
    skin_color: string
    url: string
    species: string[]
    starships: string[]
    vehicles: string[]
    films: string[]
}